+++ packages/graph-explorer/src/utils/testing/index.ts
@@ -1,4 +1,5 @@
 export * from "./DbState";
+export * from "./graphsonHelpers";
 export * from "./normalize";
 export * from "./randomData";
 export * from "./randomSchemaResponse";
